fixes bug in test model
The use of the class attribute "kinds" in doe generator was raised a NameError when used in the unittest, so got rid of it and directly defined the tuple of kinds in the generator itself. This will actually be useful to differentiate the kinds allowed for the future AtmosphereComponent and OceanComponent (which are likely to be only of type Data and Null, respectively, in a first implementation).