fixes bug for when timedomain is only one timestep long
In init, the timedelta was calculated after the elimination of the last timestamp, which was generating an IndexError when the time series covers only one timestep.

Moreover, the timedelta and period properties were realying on time and not bounds to determine their values, which, again, was a problem in the case of a one timestep long simulation.