fixes another bug for timestamps in dump files
The dump_states method was using the upper time bounds to determine the timestamp corresponding to the snapshot of the states to dump in file, but since the dumping happens between running the component for the given timestep, the lower bounds needs to be used.

Moreover, the finalise_states cannot use the dump_states method since it needs to get the timestamp for the upper bound of the last timestep (that is when the model stops), which cannot be accessed using the dump_states method (which uses the lower bounds), so finalise_states now calls update_dump_files directly.