implements proper "naked" API for initialise/run/finalise
Until now, the memory allocation for the model states was the responsibility of the model developer (i.e. allocation made in the initialise method of their beskope component class). The approach was cumbersome, and it was also problematic because there was no way to use the "science" in the initialise method while handling the memory elsewhere. This also had collateral effects on the design of the Fortran/C dummy components, to mimick what was done in Python.

Now, the initialise method is given the arrays for the model states as arguments. These arrays are initialised in the initialise_states method (a Component method). But the initialise method can now be used outside the framework, in another form of wrapper, or by another model/framework.

This commit also includes a change in the State class (related to the new way to allocate memory in initialise_states). To instantiate a State, a single array is given as argument (instead of a list/tuple of arrays before), where the first axis corresponds to the solver_history+1 (+1 for current time step). The State class will then generate a list of views of this array (effectively slicing using the first axis).

Finally, there were issues with Fortran-contiguous vs. C-contiguous arrays when reloading states from dump files namely, so now the State class takes an argument 'order' (to be used to generate numpy arrays for the states).