adds ImportError catches for C and Fortran libs
As long as they are not being used, the C/Fortran libraries don't need to exist, and if one tries to use them, it will raise a NameError because they are not defined, which is a rather informative message to track down which component is the problem.

This approaches allows for the other components (e.g. the Python version, or even the library is the other language, to be used without getting any exception raised unnecessarily.