changes axis for states divisions
Since the convention is to have T, Z, Y, X for the order of the axes in CF (i.e. time then space), it makes more sense to move the states divisions on the space side (i.e. after Z, Y, X). The first index may be used in the future for ancillary date which feature a time dimension, this way indexing [0] or indexing [-1] will give access to one or the other, with less chance for confusion.