fixes the switches in the Clock that needed an initial offset
Because the slower components need for the fastest component to provide enough information (timesteps) to cover the whole period the slower components must run for.

So an initial offset is needed. For example, if the fastest component is hourly, and another, slower, component is 3-hourly, the hourly component needs to complete its first two timesteps so that the 2-hourly component gets the fluxes that happened between the 3 hour period it needs to run for.