implements version of Interface allowing for components asynchronicity
The interface was simply a dictionary at the moment because all components were not allowed to run on different time resolutions.

This new implementation is still based on a mutable mapping paradigm, but the setitem and getitem have become more sophisticated, in order for information to be accumulated where required, and summary statistics to be produced when data from the interface is requested.

From now on, what goes to and from the interface will be referred to as a 'transfer'. This implementation puts some constraints on the transfers: a transfer is a unique exchange from a *single* component to another *single* component. As such, if a physical flux needs to be sent to two different components, they need to be named differently, because the keys 'to' and 'from' need to be unique.

A method for each transfer can be chosen between 'mean', 'sum', 'point', 'minimum' and 'maximum', and this will be applied on the overlapping period between the two timesteps of the components transferring information to one another.

The implementation determines when instantiating the Interface a list of weights for each time a transfer between components is expected (this expectation is determined by comparing the components timedomains). The weights are also determined from the timedomains, and the timesteps overlaps.

It is also to be noted that the outwards produced by a component are now not directly used to update the Interface, but a temporary dictionary at the Model level is used to buffer these outwards until all components have been (potentially) called (i.e. if the switch is 'on' for this time loop according to the Clock). This has 2 implications:
- the components effectively run with somehow 'outdated' transfers produced by their sending components;
- in turn, this means that the order in which components are called in Model does not matter anymore.
This seems to be the only way two-way coupled components do not end up in deadlocks/vicious circle, where each component is waiting for another one to complete its timestep to integrate on its own timestep.