fixes bug in special case for transfer outside framework
The value for the key arrays is meant to be a list of views into an array, so needed to set value as a list, not an array. The indexing -1 in __setitem__ was not bugging though, because the Z dimension was indexed (i.e. first array axis) instead of picking up the first element in the 'arrays', but because of the clever numpy broadcasting, the assignment was still possible. But this is a special case, so this needed fixing.