fixes bug with DataComponent not using DataSet to send to interface
When Components were streamlined (in e82af6f066249f26cba46aac2d674d6553776904), the "trick" of using the outwards of the Component being substituted as the driving data disappeared. This is now restored with this commit.

This implied that is the run method of Component, the loop through the driving data info to extract these fluxes, it did not work. But, because when Interface is set up in Model, it initialises the transfers through the Interface to arrays of zeros, when the Component requiring these fluxes will find the key in the Interface and retrieve the array of zeros rather than the data in the DataSet provided.

This was not picked up by the unittest because it currently does not check for the values being produced by the components, nor what is going through the interface.