moves check timedomain/dataset compatibility to Component
Before the check_dataset_time would be called from Model on each Component individually, which was cumbersome. Now, each time the Component is given a new TimeDomain, the availability of the data for this period in the dataset is checked automatically, and a new datasubset is generated if data is available.