allows interface to sustain its transfer values between runs
An instance of the interface was previously generated each time Model._run was called, which meant the last transfers of the interface for e.g. a spin-up run, was not used as initial conditions for the main run. Now, if an instance of the interface already exists, only a new 'setup' is carried out, and overwriting of the existing transfer values will only be done if time/space information has changed.

The passing of timedomains for each component between Model and the Components themselves has been reduced, since now the timedomain of a component is temporarily changed if required (typically for a spin-up run). The main run timedomain is stored while the spin up run(s) are happening, and will be reset when these complete.

Also, the outwards_info for a DataComponent needed to be set, otherwise the interface will not be able to carry out the transfer to the receiving Component.