moves dump to components._utils.states
Originally, I thought that we could reuse the dump functionality of the component states for the interface transfers, but they are not similar enough to recycle the existing dumping functionality, so will need to create a separate one for transfers dump. This is why the dumping functionality is repatriated in the components module, and more specifically as functions in the states file.