updates dummy components to make transfers more abstract and trackable
The inwards and outwards for the dummy components now supersedes the "real" physical ones for dummy ones. This way, tests with dummy do not need to be changed if the blueprint evolves, and also it guarantees that all components are 2-way coupled.

This required also modification of the dummy datasets, and the dummy Fortran and C versions.