moves overwriting of class attr before initialisation as Component in Null/DataC
This was necessary because otherwise these overwritten infos are not taken into account when timedomain is assigned to Component, which means that the check_dataset_time happening in the porperty timedomain would not subset the dataset to the datasubset.

Up until now, because the unittests with all combinations was going through a spinup cycle, the timedomain was reassigned to the spin-up timedomain, and this was happening after the infos were overwritten in DataComponent, so this time the datasubset would be filled in.