stores transfers even if no 'from' to receive them for dump purposes
If a real component produced this transfer, even if it is not used in the given combination of components because the receiving component is Data/NullComponent, it may be desirable to use it in the future as e.g. initial conditions. So this commit modifies the way this special case is dealt with so that 'array' and 'slices' are created so that the transfers are stored for the dump to save it.