simplifies model tests and adds checks on final state/transfer values
A Simulator class is created in test_model to be used across TestCase classes since there was a lot of redundancy. A simulator is basically a Model wrapped so that you can intialise it/spin it up/simulate with it with a minimum of arguments, just the component types and whether the components run synchronously or not.

The substitute datasets are also modified to contain the actual transfers that the 'real' components coupled together would produce. This means that e.g. a ('c', 'c', 'c') combination or a ('c', 'c', 'd') combination are meant to produce the same result (same applies to any combination of 'c's and 'd's for the component kinds).