propagate overwrite flag to _run
When using Model.resume, the overwriting of the dump files for model states is turned off in order to keep the values already in file for period before the resume start. This approach was not done for interface transfers, as the overwrite option was not propagated as far as Model._run, even though the Interface.initialise_ did offer the option. So now this is the case, and the interface dump is not overwritten anymore when a simulation is resumed.