add remapping functionality to Interface
The Interface used to hold the information at the supermesh resolution (i.e. spacedomain), which was effectively arbitrarily taken as one of the spacedomains of the components (since they could only feature the same spacedomain). Now the interface will hold the information at the resolution of the sending component and the remapping happens on "getting the value" (which avoid remapping each time information is added, as in the case of making it happen on the "setting the value").

The remapping functionality itself is pretty straighforward. With the dummy data added to the spacedomains of the components, the underlying fields of the spacedomains are used to determine the source and destination spatial configurations for the remapping. Then, it is just a matter of using cf-python to do the remapping (spherical and first order conservative for now).

Note, using cf-python for the remapping each time values are being required makes things rather slow, and in the future, an approach where the weighting for the remapping are determined once and for all at the beginning of the simulation would be better, but this may require using esmpy directly (i.e. not through cf-python, because cf-python does not given access to the weighting files). Nonetheless, this implementation gives the "correct" answer (RE: numerics), so commit as is.