deal with bug when absence of coordinate_reference attr in Field
For a RotatedLatLonGrid, the test `is_space_equal_to` was raising an AttributeError if the Field being compared to the Grid did not have a CoordinateReference. It was defaulting to None, and then trying to get attribute coordinate_conversion on NoneType.

Now it checks for presence of coordinate_reference attribute before checking further attributes coordinate_conversion and datum.

The check is also now abstracted into its own 'private' method so that it can be used by the `spans_same_region_as` method to come.