Implement remapping functionality between Components (#11)
* return deepcopies of inner fields rather than references in Space/TimeDomain

* add tolerance on equality check for space

* propagate overwrite flag to _run

When using Model.resume, the overwriting of the dump files for model states is turned off in order to keep the values already in file for period before the resume start. This approach was not done for interface transfers, as the overwrite option was not propagated as far as Model._run, even though the Interface.initialise_ did offer the option. So now this is the case, and the interface dump is not overwritten anymore when a simulation is resumed.

* add a method to Grid to check if region spanned in same as another Grid's

* add attributes to Clock/Compass to hold components' Time/SpaceDomains

* set dummy data to SpaceDomains (Grids) to allow for cf regridding

* add remapping functionality to Interface

The Interface used to hold the information at the supermesh resolution (i.e. spacedomain), which was effectively arbitrarily taken as one of the spacedomains of the components (since they could only feature the same spacedomain). Now the interface will hold the information at the resolution of the sending component and the remapping happens on "getting the value" (which avoid remapping each time information is added, as in the case of making it happen on the "setting the value").

The remapping functionality itself is pretty straighforward. With the dummy data added to the spacedomains of the components, the underlying fields of the spacedomains are used to determine the source and destination spatial configurations for the remapping. Then, it is just a matter of using cf-python to do the remapping (spherical and first order conservative for now).

Note, using cf-python for the remapping each time values are being required makes things rather slow, and in the future, an approach where the weighting for the remapping are determined once and for all at the beginning of the simulation would be better, but this may require using esmpy directly (i.e. not through cf-python, because cf-python does not given access to the weighting files). Nonetheless, this implementation gives the "correct" answer (RE: numerics), so commit as is.

* add tests with the new remapping functionality

Now the tests on the Model also test with Sync+Remap, and Async+Remap. More data files are added as required.

* deal with bug when absence of coordinate_reference attr in Field

For a RotatedLatLonGrid, the test `is_space_equal_to` was raising an AttributeError if the Field being compared to the Grid did not have a CoordinateReference. It was defaulting to None, and then trying to get attribute coordinate_conversion on NoneType.

Now it checks for presence of coordinate_reference attribute before checking further attributes coordinate_conversion and datum.

The check is also now abstracted into its own 'private' method so that it can be used by the `spans_same_region_as` method to come.

* add coordinate_reference attribute to RotateLatLonGrid

This is needed to be able to compare a Grid to self.

* add method to compare spanned regions for RotatedLatLonGrid

Needed to overwrite this method for the specific case of a rotated Grid, where the conversion parameters need to be check in addition to the extents of X Y (Z).

* add comments to locate where remap to supermesh should happen