reshape one-element long space coordinates
The latest version of `cf-python` (3.7.0) does not seem to support scalar `cf.Data` when given to populate a `cf.DimensionCoordinate` anymore.

This is only a problem for `SpaceDomain` where e.g. the Z dimension may be only one element long (but could also be the case on Y or X). In this case, coordinate dimension could have been set scalar data, and coordinate dimension bounds 1D array as a consequence. So now, after coord and coord bounds are squeezed to check their compatibility in size, if the ndim is 0, the scalar is reshaped into a 1D array, and the 1D bounds array is reshaped to a 2D one.

`TimeDomain` will always be at least two-element long (because otherwise there is no modelling timestep, so no model), so it will never be an array of ndim=0. So no need to change anything for time.

This commit also updates cf-python version required to 3.7.