simplify Model.resume() by adding argument 'tag'
The previous implementation was doing a lot of guess work to read in the latest available dump files for a given model identifier. However, beyond making the code quite messy, it also prevented from restarting from a spin up cycle (if the main run had started), or to resume an earlier spinup cycle.

By asking the user to specify a tag (i.e. 'run' if main run or 'spinup#' for a spinup cycle), this is now made possible. This also makes the code cleaner.