implement Output and OutputStream classes
The Output class is meant to hold keys attributes for a given requested component output (name, units) and a link to all output streams it must feed. It features subclasses specific to the type of output (i.e. component output/state that is already part of the interface [i.e. a transfer], bespoke component output, or bespoke component state). Each subclass implements its own __call__ method, that will be called by the component's run method (i.e. at each time iteration), and feeds the information to the relevant output streams.

An output stream gathers all the outputs of a component that feature the same requested outputting frequency. Typically, an output will be requested for a frequency equal to or a multiple of the component temporal resolution.

The OutputStream class is responsible for holding the output values until it has enough to perform the relevant aggregation methods for the output (if frequency greater than component temporal resolution). The class features output creating/writing functionality, as well as dump creating/writing functionality because the output frequency is not constrained to be equal to/a multiple of the component dumping frequency, so it needs to be handled separately from the other component dumps.

The outputs module also features a list of the supported aggregation methods as of now: mean, sum, point, min, max (including some aliases support).