add YAML constructor/respresenter/resolver for timedelta
When datetime.timedelta is used as a key (which in the case with the chosen API), the trick of using {'seconds': ###} does not work. Rather than using some bespoke strings to specify output frequency, or using the ISO 8601 for time duration, it seems better to just add support for homemade timedelta support for the YAML configuration file.

Now a timedelta can be a node in YAML directly, using the same syntax as the datetime.timedelta's API, e.g. 'timedelta(seconds=10, minutes=5)'.