fix error in timestamps of output stream
The TimeDomain defining the resolution of a Component features forward time bounds in relation to its time steps, while for the output, we need backward time bounds in relation to these time steps (and we also need to skip the first time step in the process).

This is done here simply by generating a TimeDomain for the output stream that is one time step longer than its component's. Then the time dimension is defined using the second to last component time step, and using the first to second-to-last bounds for the time bounds dimension.