move instantiation of output objects in initialise
Output objects used to be instantiated in the outputs property, so at the instantiation of Component. But in the case of a workflow with spin up + simulate runs, the output streams were reinitialised, but the link between output object-output stream was not updated. So the spinup streams were still used even if new output streams were instantiated for the simulate run.

By moving the instantiation of the oputput objects with the output streams, the link object to stream in the output object will always remain consistent. Both old objects and streams will then be dereferenced, and hence garbage collected.