fix errors in values for transfers in dummy substitude data
One value for transfer_k in subsurface. This error was not failing the tests because the impact of an error of transfer_k impacts on transfer_j which impacts on transfer_n and transfer_o. Then, transfer_n is re-used by transfer_k, but at this stage, transfer_k is from the data file, not computed by the component. And transfer_o is not used by any component. So the effect of this error dies off after a while, and the final values (which are the values checked up until now) are not impacted.