fix big when stream file/dump_file is None on resume
When using the Model resume method, the file and fump_file attributes of the output streams where not initialised (and stayed at value None). That is because overwrite=False in this case, which required an else condition both for file and dump_file, so that the location of these files can be given to the streams.