fix output format in YAML for timedomain timedelta
Since commit 1d8e509da9a208125d811820421bf47f3f9a0421, Model is expecting the custom yaml format for datetime.timedelta, but the method `to_config` of `TimeDomain` was still storing timedelta as a `dict` with a value for seconds.

The test using YAML file in test_model was not picking this up because it uses a "hand-written" YAML file, not the one generated after setting up a model. So a new test is added where a model is configured, and then its YAML file is used to configure another model, which then starts a main run and does the appropriate checks.

Note, the other test using YAML is still valid, since it guarantees that the example YAML files remain correct.