merge 'driving' and 'ancillary' under 'inputs'
Each input must now feature a 'kind' key alongside 'units' in the definition dictionary for inputs.

The 'kind' can take up to one of the three values: 'dynamic', 'static', 'climatologic'. 'dynamic' corresponds to data which features a time dimension of matching resolution with the component, 'static' does not feature a time dimension, and 'climatologic' features a time dimension, but this one represents the frequency of the climatogy covered by the data (e.g. 'monthly', 'seasonal', 'daily', etc.). If the input is of 'climatologic' kind, it must be accompanied by a 'frequency' key, and the corresponding value can be a timedelta object, or one of the two strings: 'seasonal', 'monthly', or 'day_of_year'.

If no 'kind' is provided, it is assumed to be a 'dynamic' kind. This is namely useful to provide straight compatibility of this new definition for input data for the `DataComponent`.