update dummy components using new inputs definition
'ancillary_c' for surface layer remains a 'static' kind of input (i.e. no time dimension), while 'ancillary_b' becomes a 'climatologic' kind of input (i.e. featuring a time dimension but unrelated to the component's resolution, 'monthly' for the 12 months covering a year of climatogy data), this is a new type of input data which was not supported before. All other inputs, with names featuring 'driving' in their name, are all considered a 'dynamic' kind of input.