swap determination of history for transfer in exchanger
The number of steps to be stored for a given transfer (i.e. 'history') is inferred from the depth of weights (itself determined from the relationship between the sending component time resolution and the receiving component time resolution) in order to avoid duplication of the calculations.

If the method required is 'sum', the weights area needs to be reshaped to add dimensions of size 1.

The resizing was done before the inference, so that if the method happened to be 'sum', the 'history' would take the size of the last dimension of spacedomain rather than the depth of the weights. So the resizing is moved after the inference to avoid this mistake.

In practice, all transfers are with method mean at the moment for the dummy components, so this mistake was not problematic, but could have been in the future.