upgrade Exchanger to support one-to-many transfers
The Exchanger used to inherit from collections.abc.MutableMapping, because a transfer was only allowed for one component-to-one component. However, OW's 'water_level' will need to be transferred to both SL and SS. And this is likely to happen again when considering other cycles (e.g. nutrient, carbon, energy).

In moving to a one-to-many support, the intention of this new implementation is to avoid the duplication of information, i.e. if SL needs a history of 1 and SS needs a history of 3, it should store a history of 3, and SL will only use the last value while SS will use all 3 values.

The direct MutableMapping approach is not sustainable in these conditions, because the __getitem__ will not be given enough information about the requesting component to know which history to provide. An alternative would have been to rename the transfer by adding the name of the component in the name of the transfer, but then the name defined in the component's interface would have to be complemented using its category.

Instead, the MutableMapping inheritence is dropped, and new methods get_transfer, set_transfer, and update_transfers are implemented. Additional information is added in the *transfers* dictionary specific to each requesting component: 'remap', 'history', 'weights', and 'iter' now exist for each requesting component, and are stored in *transfers* using the component's category as the key.

`__getitem__` becoming `get_transfer` requires only minimal modifications: it just needs to collect 'history' to know which 'slices' to return.

`__setitem__` becoming set_transfer requires no modification.

`update` cannot be directly inherited 'for free' as was the case with a MutableMapping, and `update_transfers` needs to be implemented: it simply iterates over the transfers and call `set_transfer` for each.

Note, in the definition of the interface in the components (i.e. inwards/outwards), the 'to' key in *outwards_info* now needs to feature a sequence as value, even if only one component is the destination.