fix mistake in weights for method 'sum'
The _calculate_weights method of Exchanger returns an array of integers (which is better than floating point numbers for precision), for a weighted 'mean' this is fine, but for a weighted 'sum', the values need to sum to one.

This could be done later in __get_item__, but this would mean scaling the weights do that they sum to one many times, so it seems better to do it when setting up Exchanger.