replace timedelta by integer support for climatologic frequency
Timedelta assumes a gregorian calendar, which is not flexible, and risks behing inconsistent with the calendar of the component. So for now, there will only be a support for a custom integer value, corresponding to the number of sub-periods for which climatologic values are required in the calendar year. It is imperfect as well, but at least it drops the explicit assumption of a specific calendar.