Gather 'driving'/'ancillary' data into a single 'inputs' item (#15)
* merge 'driving' and 'ancillary' under 'inputs'

Each input must now feature a 'kind' key alongside 'units' in the definition dictionary for inputs.

The 'kind' can take up to one of the three values: 'dynamic', 'static', 'climatologic'. 'dynamic' corresponds to data which features a time dimension of matching resolution with the component, 'static' does not feature a time dimension, and 'climatologic' features a time dimension, but this one represents the frequency of the climatogy covered by the data (e.g. 'monthly', 'seasonal', 'daily', etc.). If the input is of 'climatologic' kind, it must be accompanied by a 'frequency' key, and the corresponding value can be a timedelta object, or one of the two strings: 'seasonal', 'monthly', or 'day_of_year'.

If no 'kind' is provided, it is assumed to be a 'dynamic' kind. This is namely useful to provide straight compatibility of this new definition for input data for the `DataComponent`.

* add check on component subclass definition

Model developers must populate the various class attributes (dictionaries) 'inputs_info', 'parameters_info', etc.

Up until now, there was no check on the presence of sufficient information for the proper functioning of their component. E.g. all items in these dictionaries must feature 'units', and more specific cases exist for 'inputs_info' now.

So this commit adds a new method for `Component` class, which does those checks in the component's definition.

* update dummy components using new inputs definition

'ancillary_c' for surface layer remains a 'static' kind of input (i.e. no time dimension), while 'ancillary_b' becomes a 'climatologic' kind of input (i.e. featuring a time dimension but unrelated to the component's resolution, 'monthly' for the 12 months covering a year of climatogy data), this is a new type of input data which was not supported before. All other inputs, with names featuring 'driving' in their name, are all considered a 'dynamic' kind of input.

* rename NC variable for X/Y/Z bounds in dummy data

The netCDF variable names for the altitude/latitude/longitude bounds in the file created with `cf-python` were arbitrary (e.g. 'bounds', 'bounds1', 'bounds2'), so this commit renames them with more explicit names altitude_bounds, etc.

* update 'ancillary_b' in dummy data to make it 'climatologic'

* fix typo in class attribute in `DataComponent`

* update comments in `Component` to remove mentions of driving/ancil

* remove unused packages

* update sciencish components using new inputs definition

* add details for climatologic frequency in docstring

* fix typo in docstring

* replace timedelta by integer support for climatologic frequency

Timedelta assumes a gregorian calendar, which is not flexible, and risks behing inconsistent with the calendar of the component. So for now, there will only be a support for a custom integer value, corresponding to the number of sub-periods for which climatologic values are required in the calendar year. It is imperfect as well, but at least it drops the explicit assumption of a specific calendar.