add current_datetime attr to Component
Many models will want access to the current simulation datetime, e.g. to determine the month the timestep is in, or to determine the day of the year, etc. So a new property is given to `Component`, that is intially set to the start of the first timestep, and will be updted each time the 'run' method is called.