rename attr to timedelta_in_seconds in Component
'timestepinseconds' was a bit ambiguous, because it could be understood as:
- the gap between two consecutive points in time in the temporal resolution (which it is)
- the gap between the beginning of the simulation and the current point the simulation is at, like the way time is stored as a floating point in a netCDF file (which it is not)

'timedelta_in_seconds' sounds less ambiguous to me, because it is close to the "delta t" a numerical modeller would use in an equation. Also, it uses the same terminology as the other property 'timedelta' which returns a timedelta object, so this seems more consistent.