drop bounds in LatLonGrid comparison with field if field has none
Quoting CF-1.8: "If bounds are not provided, an application might reasonably assume the gridpoints to be at the centers of the cells, but we do not require that in this standard."

Since LatLonGrid is required to have bounds for conservative remapping with cf-python/ESMF, the comparison with a field for compatibility was more restrictive than what the CF standard requires.

So, if a field with no bounds is compared with LatLonGrid, the new implementation of 'is_space_equal_to' momentarily removes the bounds of the LatLonGrid before comparing the two constructs, and then appends it back to LatLonGrid. This is more straightforward than comparing each sub-construct separately, simply to avoid the bounds comparison.

If a field has bounds, the comparison remains unchanged compared to before, and everything making up the dimension_coordinate will be compared.