add subset in space on data fields using spacedomain
Similar to what was done for time already, the spacedomain is used to subset in space the data field given to the component, to select the region of interest, if the data region is greater than the simulation region.

This functionality is only supported for a `Grid`, because this is the only spatial configuration where the subset in space is straightfoward using X and Y coordinates.