Upgrade from one-to-one to one-to-many transfer support (#19)
* merge 'driving' and 'ancillary' under 'inputs'

Each input must now feature a 'kind' key alongside 'units' in the definition dictionary for inputs.

The 'kind' can take up to one of the three values: 'dynamic', 'static', 'climatologic'. 'dynamic' corresponds to data which features a time dimension of matching resolution with the component, 'static' does not feature a time dimension, and 'climatologic' features a time dimension, but this one represents the frequency of the climatogy covered by the data (e.g. 'monthly', 'seasonal', 'daily', etc.). If the input is of 'climatologic' kind, it must be accompanied by a 'frequency' key, and the corresponding value can be a timedelta object, or one of the two strings: 'seasonal', 'monthly', or 'day_of_year'.

If no 'kind' is provided, it is assumed to be a 'dynamic' kind. This is namely useful to provide straight compatibility of this new definition for input data for the `DataComponent`.

* add check on component subclass definition

Model developers must populate the various class attributes (dictionaries) 'inputs_info', 'parameters_info', etc.

Up until now, there was no check on the presence of sufficient information for the proper functioning of their component. E.g. all items in these dictionaries must feature 'units', and more specific cases exist for 'inputs_info' now.

So this commit adds a new method for `Component` class, which does those checks in the component's definition.

* update dummy components using new inputs definition

'ancillary_c' for surface layer remains a 'static' kind of input (i.e. no time dimension), while 'ancillary_b' becomes a 'climatologic' kind of input (i.e. featuring a time dimension but unrelated to the component's resolution, 'monthly' for the 12 months covering a year of climatogy data), this is a new type of input data which was not supported before. All other inputs, with names featuring 'driving' in their name, are all considered a 'dynamic' kind of input.

* rename NC variable for X/Y/Z bounds in dummy data

The netCDF variable names for the altitude/latitude/longitude bounds in the file created with `cf-python` were arbitrary (e.g. 'bounds', 'bounds1', 'bounds2'), so this commit renames them with more explicit names altitude_bounds, etc.

* update 'ancillary_b' in dummy data to make it 'climatologic'

* fix typo in class attribute in `DataComponent`

* update comments in `Component` to remove mentions of driving/ancil

* remove unused packages

* rename Interface to Exchanger

This is a more descriptive name for the class, and it is useful to distinguish between the interface and the exchanger:
- the 'interface' is the collection of transfers between components (i.e. what scientific information is being exchanged);
- the 'exchanger' is the enabler for these transfers to happen (i.e. it is the computational implementation allowing for the scientific information to be exchanged).

Note, for the outputs, the class `InterfaceOutput` is not renamed because it refers to the former rather than to the latter of the two concepts above.

* update sciencish components using new inputs definition

* swap determination of history for transfer in exchanger

The number of steps to be stored for a given transfer (i.e. 'history') is inferred from the depth of weights (itself determined from the relationship between the sending component time resolution and the receiving component time resolution) in order to avoid duplication of the calculations.

If the method required is 'sum', the weights area needs to be reshaped to add dimensions of size 1.

The resizing was done before the inference, so that if the method happened to be 'sum', the 'history' would take the size of the last dimension of spacedomain rather than the depth of the weights. So the resizing is moved after the inference to avoid this mistake.

In practice, all transfers are with method mean at the moment for the dummy components, so this mistake was not problematic, but could have been in the future.

* upgrade Exchanger to support one-to-many transfers

The Exchanger used to inherit from collections.abc.MutableMapping, because a transfer was only allowed for one component-to-one component. However, OW's 'water_level' will need to be transferred to both SL and SS. And this is likely to happen again when considering other cycles (e.g. nutrient, carbon, energy).

In moving to a one-to-many support, the intention of this new implementation is to avoid the duplication of information, i.e. if SL needs a history of 1 and SS needs a history of 3, it should store a history of 3, and SL will only use the last value while SS will use all 3 values.

The direct MutableMapping approach is not sustainable in these conditions, because the __getitem__ will not be given enough information about the requesting component to know which history to provide. An alternative would have been to rename the transfer by adding the name of the component in the name of the transfer, but then the name defined in the component's interface would have to be complemented using its category.

Instead, the MutableMapping inheritence is dropped, and new methods get_transfer, set_transfer, and update_transfers are implemented. Additional information is added in the *transfers* dictionary specific to each requesting component: 'remap', 'history', 'weights', and 'iter' now exist for each requesting component, and are stored in *transfers* using the component's category as the key.

`__getitem__` becoming `get_transfer` requires only minimal modifications: it just needs to collect 'history' to know which 'slices' to return.

`__setitem__` becoming set_transfer requires no modification.

`update` cannot be directly inherited 'for free' as was the case with a MutableMapping, and `update_transfers` needs to be implemented: it simply iterates over the transfers and call `set_transfer` for each.

Note, in the definition of the interface in the components (i.e. inwards/outwards), the 'to' key in *outwards_info* now needs to feature a sequence as value, even if only one component is the destination.

* update dummy interface definition to comply with new exchanger

* add details for climatologic frequency in docstring

* fix typo in docstring

* update dummy interface to add a one-to-many transfer

OW's 'transfer_n' is now sent to both SL and SS. SL uses it in the calculation of its 'output_x'.

The values are updated in test_outputs to reflect this change.

* refactor tests structure to follow package structure

* add doctest for Exchanger._calculate_weights

* fix mistake in weights for method 'sum'

The _calculate_weights method of Exchanger returns an array of integers (which is better than floating point numbers for precision), for a weighted 'mean' this is fine, but for a weighted 'sum', the values need to sum to one.

This could be done later in __get_item__, but this would mean scaling the weights do that they sum to one many times, so it seems better to do it when setting up Exchanger.

* remove redundant resetting of clock/compass in Exchanger

* remove history setting for special case

Since there is a one to many approach for the transfer, the history is determined as the maximum of the history of each transfer, so no need to set it so early.

* replace timedelta by integer support for climatologic frequency

Timedelta assumes a gregorian calendar, which is not flexible, and risks behing inconsistent with the calendar of the component. So for now, there will only be a support for a custom integer value, corresponding to the number of sub-periods for which climatologic values are required in the calendar year. It is imperfect as well, but at least it drops the explicit assumption of a specific calendar.

* remove unused import